refactor: split nat_gateway.go by concern
Pure code move with no behavior change. Splits the ~900-line
nat_gateway.go into per-concern files along the existing service
interface seams:

- nat_gateway.go            CRUD, design/buy lifecycle, sessions,
                            telemetry, shared doJSON helper
- nat_gateway_validation.go general structural validators + shared
                            Err vars
- nat_gateway_packet_filters.go
- nat_gateway_prefix_lists.go
- nat_gateway_diagnostics.go

Lands first so follow-up work (matrix-aware validator, ESD-1203)
lives in clean files and review diffs stay focused.

Refs: ESD-1202
