+++ nat_gateway.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
-	"slices"
 	"strconv"
 	"time"
 )
@@ -122,94 +121,6 @@ type GetNATGatewayTelemetryRequest struct {
 	From       *time.Time // Start time. Mutually exclusive with Days.
 	To         *time.Time // End time. Mutually exclusive with Days.
 	Days       *int32     // Number of days of telemetry (1-180). Mutually exclusive with From/To.
-}
-
-// ErrNATGatewayProductUIDRequired is returned when a ProductUID is not provided.
-var ErrNATGatewayProductUIDRequired = errors.New("product UID is required")
-
-// ErrNATGatewayTelemetryTypesRequired is returned when no telemetry types are provided.
-var ErrNATGatewayTelemetryTypesRequired = errors.New("at least one telemetry type is required")
-
-// ErrNATGatewayTelemetryTimeExclusive is returned when both Days and From/To are provided.
-var ErrNATGatewayTelemetryTimeExclusive = errors.New("days and from/to are mutually exclusive")
-
-// ErrNATGatewayTelemetryDaysOutOfRange is returned when Days is not between 1 and 180.
-var ErrNATGatewayTelemetryDaysOutOfRange = errors.New("days must be between 1 and 180")
-
-// ErrNATGatewayTelemetryFromToIncomplete is returned when only one of From/To is provided.
-var ErrNATGatewayTelemetryFromToIncomplete = errors.New("both from and to must be provided together")
-
-// ErrNATGatewayProductNameRequired is returned when a ProductName is not provided.
-var ErrNATGatewayProductNameRequired = errors.New("product name is required")
-
-// ErrNATGatewayLocationIDRequired is returned when a LocationID is not provided or is invalid.
-var ErrNATGatewayLocationIDRequired = errors.New("location ID must be greater than 0")
-
-// ErrNATGatewaySpeedRequired is returned when a Speed is not provided or is invalid.
-var ErrNATGatewaySpeedRequired = errors.New("speed must be greater than 0")
-
-// ErrNATGatewayInvalidTerm is returned when a Term is not a valid contract term.
-var ErrNATGatewayInvalidTerm = errors.New("term must be one of: 1, 12, 24, 36, 48, 60")
-
-// Packet filter validation errors.
-var (
-	ErrNATGatewayPacketFilterIDRequired       = errors.New("packet filter ID must be greater than 0")
-	ErrNATGatewayPacketFilterDescriptionEmpty = errors.New("packet filter description is required")
-	ErrNATGatewayPacketFilterEntriesEmpty     = errors.New("packet filter requires at least one entry")
-)
-
-// Prefix list validation errors.
-var (
-	ErrNATGatewayPrefixListIDRequired         = errors.New("prefix list ID must be greater than 0")
-	ErrNATGatewayPrefixListDescriptionEmpty   = errors.New("prefix list description is required")
-	ErrNATGatewayPrefixListAddressFamilyEmpty = errors.New("prefix list addressFamily is required")
-	ErrNATGatewayPrefixListEntriesEmpty       = errors.New("prefix list requires at least one entry")
-	ErrNATGatewayPrefixListEmptyResponse      = errors.New("API returned a 2xx response with an empty prefix list payload")
-)
-
-// Diagnostics validation errors.
-var (
-	ErrNATGatewayDiagnosticsPeerIPRequired   = errors.New("BGP neighbor diagnostics require a peer IP address")
-	ErrNATGatewayDiagnosticsDirectionInvalid = errors.New("BGP neighbor diagnostics require a direction (RECEIVED or ADVERTISED)")
-	ErrNATGatewayDiagnosticsOperationEmpty   = errors.New("operation ID is required")
-	ErrNATGatewayDiagnosticsTimeout          = errors.New("timed out waiting for diagnostics operation to complete")
-)
-
-// validateCreateNATGatewayRequest validates the request parameters for creating a NAT Gateway.
-func validateCreateNATGatewayRequest(req *CreateNATGatewayRequest) error {
-	if req.ProductName == "" {
-		return ErrNATGatewayProductNameRequired
-	}
-	if req.LocationID < 1 {
-		return ErrNATGatewayLocationIDRequired
-	}
-	if req.Speed < 1 {
-		return ErrNATGatewaySpeedRequired
-	}
-	if !slices.Contains(VALID_CONTRACT_TERMS, req.Term) {
-		return ErrNATGatewayInvalidTerm
-	}
-	return nil
-}
-
-// validateUpdateNATGatewayRequest validates the request parameters for updating a NAT Gateway.
-func validateUpdateNATGatewayRequest(req *UpdateNATGatewayRequest) error {
-	if req.ProductUID == "" {
-		return ErrNATGatewayProductUIDRequired
-	}
-	if req.ProductName == "" {
-		return ErrNATGatewayProductNameRequired
-	}
-	if req.LocationID < 1 {
-		return ErrNATGatewayLocationIDRequired
-	}
-	if req.Speed < 1 {
-		return ErrNATGatewaySpeedRequired
-	}
-	if !slices.Contains(VALID_CONTRACT_TERMS, req.Term) {
-		return ErrNATGatewayInvalidTerm
-	}
-	return nil
 }
 
 // CreateNATGateway creates a new NAT Gateway resource.
@@ -428,26 +339,6 @@ func (svc *NATGatewayServiceOp) postNetworkDesign(ctx context.Context, path, pro
 	return nil
 }
 
-// validateGetNATGatewayTelemetryRequest validates the request parameters.
-func validateGetNATGatewayTelemetryRequest(req *GetNATGatewayTelemetryRequest) error {
-	if req.ProductUID == "" {
-		return ErrNATGatewayProductUIDRequired
-	}
-	if len(req.Types) == 0 {
-		return ErrNATGatewayTelemetryTypesRequired
-	}
-	if req.Days != nil && (req.From != nil || req.To != nil) {
-		return ErrNATGatewayTelemetryTimeExclusive
-	}
-	if req.Days != nil && (*req.Days < 1 || *req.Days > 180) {
-		return ErrNATGatewayTelemetryDaysOutOfRange
-	}
-	if (req.From != nil) != (req.To != nil) {
-		return ErrNATGatewayTelemetryFromToIncomplete
-	}
-	return nil
-}
-
 // ListNATGatewaySessions returns the speed/session-count availability matrix for NAT Gateways.
 func (svc *NATGatewayServiceOp) ListNATGatewaySessions(ctx context.Context) ([]*NATGatewaySession, error) {
 	path := "/v3/products/nat_gateways/sessions"
@@ -506,380 +397,6 @@ func (svc *NATGatewayServiceOp) GetNATGatewayTelemetry(ctx context.Context, req 
 	}
 	return telemetryResp, nil
 }
-
-// --- Packet filters -------------------------------------------------------
-
-func validateNATGatewayPacketFilterRequest(req *NATGatewayPacketFilterRequest) error {
-	if req == nil || req.Description == "" {
-		return ErrNATGatewayPacketFilterDescriptionEmpty
-	}
-	if len(req.Entries) == 0 {
-		return ErrNATGatewayPacketFilterEntriesEmpty
-	}
-	return nil
-}
-
-// ListNATGatewayPacketFilters returns all packet filter summaries for a NAT Gateway.
-func (svc *NATGatewayServiceOp) ListNATGatewayPacketFilters(ctx context.Context, productUID string) ([]*NATGatewayPacketFilterSummary, error) {
-	if productUID == "" {
-		return nil, ErrNATGatewayProductUIDRequired
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/packet_filter_summaries", url.PathEscape(productUID))
-	var envelope natGatewayPacketFilterSummariesResponse
-	if err := svc.doJSON(ctx, http.MethodGet, path, nil, &envelope); err != nil {
-		return nil, err
-	}
-	return envelope.Data, nil
-}
-
-// CreateNATGatewayPacketFilter creates a new packet filter on a NAT Gateway.
-func (svc *NATGatewayServiceOp) CreateNATGatewayPacketFilter(ctx context.Context, productUID string, req *NATGatewayPacketFilterRequest) (*NATGatewayPacketFilter, error) {
-	if productUID == "" {
-		return nil, ErrNATGatewayProductUIDRequired
-	}
-	if err := validateNATGatewayPacketFilterRequest(req); err != nil {
-		return nil, err
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/packet_filters", url.PathEscape(productUID))
-	var envelope natGatewayPacketFilterResponse
-	if err := svc.doJSON(ctx, http.MethodPost, path, req, &envelope); err != nil {
-		return nil, err
-	}
-	return envelope.Data, nil
-}
-
-// GetNATGatewayPacketFilter returns a packet filter by its numeric ID.
-func (svc *NATGatewayServiceOp) GetNATGatewayPacketFilter(ctx context.Context, productUID string, packetFilterID int) (*NATGatewayPacketFilter, error) {
-	if productUID == "" {
-		return nil, ErrNATGatewayProductUIDRequired
-	}
-	if packetFilterID < 1 {
-		return nil, ErrNATGatewayPacketFilterIDRequired
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/packet_filters/%d", url.PathEscape(productUID), packetFilterID)
-	var envelope natGatewayPacketFilterResponse
-	if err := svc.doJSON(ctx, http.MethodGet, path, nil, &envelope); err != nil {
-		return nil, err
-	}
-	return envelope.Data, nil
-}
-
-// UpdateNATGatewayPacketFilter replaces a packet filter's description and entries.
-func (svc *NATGatewayServiceOp) UpdateNATGatewayPacketFilter(ctx context.Context, productUID string, packetFilterID int, req *NATGatewayPacketFilterRequest) (*NATGatewayPacketFilter, error) {
-	if productUID == "" {
-		return nil, ErrNATGatewayProductUIDRequired
-	}
-	if packetFilterID < 1 {
-		return nil, ErrNATGatewayPacketFilterIDRequired
-	}
-	if err := validateNATGatewayPacketFilterRequest(req); err != nil {
-		return nil, err
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/packet_filters/%d", url.PathEscape(productUID), packetFilterID)
-	var envelope natGatewayPacketFilterResponse
-	if err := svc.doJSON(ctx, http.MethodPut, path, req, &envelope); err != nil {
-		return nil, err
-	}
-	return envelope.Data, nil
-}
-
-// DeleteNATGatewayPacketFilter removes a packet filter from a NAT Gateway.
-func (svc *NATGatewayServiceOp) DeleteNATGatewayPacketFilter(ctx context.Context, productUID string, packetFilterID int) error {
-	if productUID == "" {
-		return ErrNATGatewayProductUIDRequired
-	}
-	if packetFilterID < 1 {
-		return ErrNATGatewayPacketFilterIDRequired
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/packet_filters/%d", url.PathEscape(productUID), packetFilterID)
-	return svc.doJSON(ctx, http.MethodDelete, path, nil, nil)
-}
-
-// --- Prefix lists ---------------------------------------------------------
-
-func validateNATGatewayPrefixList(req *NATGatewayPrefixList) error {
-	if req == nil || req.Description == "" {
-		return ErrNATGatewayPrefixListDescriptionEmpty
-	}
-	if req.AddressFamily == "" {
-		return ErrNATGatewayPrefixListAddressFamilyEmpty
-	}
-	if len(req.Entries) == 0 {
-		return ErrNATGatewayPrefixListEntriesEmpty
-	}
-	return nil
-}
-
-// ListNATGatewayPrefixLists returns all prefix list summaries for a NAT Gateway.
-func (svc *NATGatewayServiceOp) ListNATGatewayPrefixLists(ctx context.Context, productUID string) ([]*NATGatewayPrefixListSummary, error) {
-	if productUID == "" {
-		return nil, ErrNATGatewayProductUIDRequired
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/prefix_list_summaries", url.PathEscape(productUID))
-	var envelope natGatewayPrefixListSummariesResponse
-	if err := svc.doJSON(ctx, http.MethodGet, path, nil, &envelope); err != nil {
-		return nil, err
-	}
-	return envelope.Data, nil
-}
-
-// CreateNATGatewayPrefixList creates a new prefix list on a NAT Gateway.
-func (svc *NATGatewayServiceOp) CreateNATGatewayPrefixList(ctx context.Context, productUID string, req *NATGatewayPrefixList) (*NATGatewayPrefixList, error) {
-	if productUID == "" {
-		return nil, ErrNATGatewayProductUIDRequired
-	}
-	if err := validateNATGatewayPrefixList(req); err != nil {
-		return nil, err
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/prefix_lists", url.PathEscape(productUID))
-	var envelope natGatewayPrefixListResponse
-	if err := svc.doJSON(ctx, http.MethodPost, path, req.toAPI(), &envelope); err != nil {
-		return nil, err
-	}
-	if envelope.Data == nil {
-		return nil, ErrNATGatewayPrefixListEmptyResponse
-	}
-	return envelope.Data.toPrefixList()
-}
-
-// GetNATGatewayPrefixList returns a prefix list by its numeric ID.
-func (svc *NATGatewayServiceOp) GetNATGatewayPrefixList(ctx context.Context, productUID string, prefixListID int) (*NATGatewayPrefixList, error) {
-	if productUID == "" {
-		return nil, ErrNATGatewayProductUIDRequired
-	}
-	if prefixListID < 1 {
-		return nil, ErrNATGatewayPrefixListIDRequired
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/prefix_lists/%d", url.PathEscape(productUID), prefixListID)
-	var envelope natGatewayPrefixListResponse
-	if err := svc.doJSON(ctx, http.MethodGet, path, nil, &envelope); err != nil {
-		return nil, err
-	}
-	if envelope.Data == nil {
-		return nil, ErrNATGatewayPrefixListEmptyResponse
-	}
-	return envelope.Data.toPrefixList()
-}
-
-// UpdateNATGatewayPrefixList replaces a prefix list's description, address family, and entries.
-func (svc *NATGatewayServiceOp) UpdateNATGatewayPrefixList(ctx context.Context, productUID string, prefixListID int, req *NATGatewayPrefixList) (*NATGatewayPrefixList, error) {
-	if productUID == "" {
-		return nil, ErrNATGatewayProductUIDRequired
-	}
-	if prefixListID < 1 {
-		return nil, ErrNATGatewayPrefixListIDRequired
-	}
-	if err := validateNATGatewayPrefixList(req); err != nil {
-		return nil, err
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/prefix_lists/%d", url.PathEscape(productUID), prefixListID)
-	var envelope natGatewayPrefixListResponse
-	if err := svc.doJSON(ctx, http.MethodPut, path, req.toAPI(), &envelope); err != nil {
-		return nil, err
-	}
-	if envelope.Data == nil {
-		return nil, ErrNATGatewayPrefixListEmptyResponse
-	}
-	return envelope.Data.toPrefixList()
-}
-
-// DeleteNATGatewayPrefixList removes a prefix list from a NAT Gateway.
-func (svc *NATGatewayServiceOp) DeleteNATGatewayPrefixList(ctx context.Context, productUID string, prefixListID int) error {
-	if productUID == "" {
-		return ErrNATGatewayProductUIDRequired
-	}
-	if prefixListID < 1 {
-		return ErrNATGatewayPrefixListIDRequired
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/prefix_lists/%d", url.PathEscape(productUID), prefixListID)
-	return svc.doJSON(ctx, http.MethodDelete, path, nil, nil)
-}
-
-// --- Diagnostics ----------------------------------------------------------
-
-// diagnosticsPollInterval and diagnosticsPollTimeout are the defaults used by
-// the convenience diagnostics wrappers.
-const (
-	diagnosticsPollInitialDelay = 2 * time.Second
-	diagnosticsPollInterval     = 3 * time.Second
-	diagnosticsPollTimeout      = 60 * time.Second
-)
-
-// ListNATGatewayIPRoutesAsync submits an IP routes diagnostics request.
-func (svc *NATGatewayServiceOp) ListNATGatewayIPRoutesAsync(ctx context.Context, productUID, ipAddress string) (string, error) {
-	if productUID == "" {
-		return "", ErrNATGatewayProductUIDRequired
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/diagnostics/routes/ip", url.PathEscape(productUID))
-	if ipAddress != "" {
-		params := url.Values{}
-		params.Set("ip_address", ipAddress)
-		path = path + "?" + params.Encode()
-	}
-	var envelope natGatewayDiagnosticsAsyncResponse
-	if err := svc.doJSON(ctx, http.MethodGet, path, nil, &envelope); err != nil {
-		return "", err
-	}
-	return envelope.Data, nil
-}
-
-// ListNATGatewayBGPRoutesAsync submits a BGP routes diagnostics request.
-func (svc *NATGatewayServiceOp) ListNATGatewayBGPRoutesAsync(ctx context.Context, productUID, ipAddress string) (string, error) {
-	if productUID == "" {
-		return "", ErrNATGatewayProductUIDRequired
-	}
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/diagnostics/routes/bgp", url.PathEscape(productUID))
-	if ipAddress != "" {
-		params := url.Values{}
-		params.Set("ip_address", ipAddress)
-		path = path + "?" + params.Encode()
-	}
-	var envelope natGatewayDiagnosticsAsyncResponse
-	if err := svc.doJSON(ctx, http.MethodGet, path, nil, &envelope); err != nil {
-		return "", err
-	}
-	return envelope.Data, nil
-}
-
-// ListNATGatewayBGPNeighborRoutesAsync submits a BGP neighbor routes diagnostics request.
-func (svc *NATGatewayServiceOp) ListNATGatewayBGPNeighborRoutesAsync(ctx context.Context, req *NATGatewayBGPNeighborRoutesRequest) (string, error) {
-	if req == nil || req.ProductUID == "" {
-		return "", ErrNATGatewayProductUIDRequired
-	}
-	if req.PeerIPAddress == "" {
-		return "", ErrNATGatewayDiagnosticsPeerIPRequired
-	}
-	if req.Direction != BGPRouteDirectionReceived && req.Direction != BGPRouteDirectionAdvertised {
-		return "", ErrNATGatewayDiagnosticsDirectionInvalid
-	}
-	params := url.Values{}
-	params.Set("direction", req.Direction)
-	params.Set("peer_ip_address", req.PeerIPAddress)
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/diagnostics/routes/bgp/neighbor?%s", url.PathEscape(req.ProductUID), params.Encode())
-	var envelope natGatewayDiagnosticsAsyncResponse
-	if err := svc.doJSON(ctx, http.MethodGet, path, nil, &envelope); err != nil {
-		return "", err
-	}
-	return envelope.Data, nil
-}
-
-// GetNATGatewayDiagnosticsRoutes fetches the routes for a prior async request.
-func (svc *NATGatewayServiceOp) GetNATGatewayDiagnosticsRoutes(ctx context.Context, productUID, operationID string) ([]*NATGatewayRoute, error) {
-	if productUID == "" {
-		return nil, ErrNATGatewayProductUIDRequired
-	}
-	if operationID == "" {
-		return nil, ErrNATGatewayDiagnosticsOperationEmpty
-	}
-	params := url.Values{}
-	params.Set("operationId", operationID)
-	path := fmt.Sprintf("/v3/products/nat_gateways/%s/diagnostics/routes/operation?%s", url.PathEscape(productUID), params.Encode())
-	var envelope natGatewayDiagnosticsRoutesResponse
-	if err := svc.doJSON(ctx, http.MethodGet, path, nil, &envelope); err != nil {
-		return nil, err
-	}
-	return envelope.Data, nil
-}
-
-// pollDiagnosticsRoutes polls GetNATGatewayDiagnosticsRoutes until the
-// operation returns a non-empty result, the SDK-managed
-// diagnosticsPollTimeout elapses, or the caller's context is cancelled.
-// Empty responses are treated as "still processing".
-func (svc *NATGatewayServiceOp) pollDiagnosticsRoutes(ctx context.Context, productUID, operationID string) ([]*NATGatewayRoute, error) {
-	pollCtx, cancel := context.WithTimeout(ctx, diagnosticsPollTimeout)
-	defer cancel()
-	// pollDoneErr returns ctx.Err() when the caller's context is the one that
-	// fired (cancellation or caller-imposed deadline) and
-	// ErrNATGatewayDiagnosticsTimeout when the SDK-managed
-	// diagnosticsPollTimeout is what elapsed. This lets callers tell
-	// "my deadline hit" from "the diagnostics op never completed".
-	pollDoneErr := func() error {
-		if err := ctx.Err(); err != nil {
-			return err
-		}
-		return ErrNATGatewayDiagnosticsTimeout
-	}
-	select {
-	case <-pollCtx.Done():
-		return nil, pollDoneErr()
-	case <-time.After(diagnosticsPollInitialDelay):
-	}
-	ticker := time.NewTicker(diagnosticsPollInterval)
-	defer ticker.Stop()
-	for {
-		routes, err := svc.GetNATGatewayDiagnosticsRoutes(pollCtx, productUID, operationID)
-		if err != nil {
-			return nil, err
-		}
-		if len(routes) > 0 {
-			return routes, nil
-		}
-		select {
-		case <-pollCtx.Done():
-			return nil, pollDoneErr()
-		case <-ticker.C:
-		}
-	}
-}
-
-// ListNATGatewayIPRoutes submits an IP routes request and polls until results are available.
-func (svc *NATGatewayServiceOp) ListNATGatewayIPRoutes(ctx context.Context, productUID, ipAddress string) ([]*NATGatewayIPRoute, error) {
-	opID, err := svc.ListNATGatewayIPRoutesAsync(ctx, productUID, ipAddress)
-	if err != nil {
-		return nil, err
-	}
-	routes, err := svc.pollDiagnosticsRoutes(ctx, productUID, opID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]*NATGatewayIPRoute, 0, len(routes))
-	for _, r := range routes {
-		if r.IP != nil {
-			out = append(out, r.IP)
-		}
-	}
-	return out, nil
-}
-
-// ListNATGatewayBGPRoutes submits a BGP routes request and polls until results are available.
-func (svc *NATGatewayServiceOp) ListNATGatewayBGPRoutes(ctx context.Context, productUID, ipAddress string) ([]*NATGatewayBGPRoute, error) {
-	opID, err := svc.ListNATGatewayBGPRoutesAsync(ctx, productUID, ipAddress)
-	if err != nil {
-		return nil, err
-	}
-	routes, err := svc.pollDiagnosticsRoutes(ctx, productUID, opID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]*NATGatewayBGPRoute, 0, len(routes))
-	for _, r := range routes {
-		if r.BGP != nil {
-			out = append(out, r.BGP)
-		}
-	}
-	return out, nil
-}
-
-// ListNATGatewayBGPNeighborRoutes submits a BGP neighbor routes request and polls for results.
-func (svc *NATGatewayServiceOp) ListNATGatewayBGPNeighborRoutes(ctx context.Context, req *NATGatewayBGPNeighborRoutesRequest) ([]*NATGatewayBGPRoute, error) {
-	opID, err := svc.ListNATGatewayBGPNeighborRoutesAsync(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	routes, err := svc.pollDiagnosticsRoutes(ctx, req.ProductUID, opID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]*NATGatewayBGPRoute, 0, len(routes))
-	for _, r := range routes {
-		if r.BGP != nil {
-			out = append(out, r.BGP)
-		}
-	}
-	return out, nil
-}
-
-// --- shared helper --------------------------------------------------------
 
 // doJSON sends a JSON request and decodes the response into out (or discards
 // the body if out is nil). It centralises the NewRequest/Do/Unmarshal dance
